Optimize rst_to_html function for speed
