+++ lib/galaxy/util/__init__.py
@@ -7,7 +7,6 @@ import binascii
 import codecs
 import collections
 import errno
-import functools
 import importlib
 import itertools
 import json
@@ -87,16 +86,6 @@ except ImportError:
         ElementTree,
     )
 
-try:
-    import docutils.core
-    import docutils.io
-    import docutils.utils
-    import docutils.writers.html4css1
-except ImportError:
-    docutils_core = None  # type: ignore[assignment]
-    docutils_html4css1 = None  # type: ignore[assignment]
-    docutils_utils = None  # type: ignore[assignment]
-
 from .custom_logging import get_logger
 from .inflection import Inflector
 from .path import (  # noqa: F401
@@ -105,6 +94,7 @@ from .path import (  # noqa: F401
     safe_relpath,
     StrPath,
 )
+from .rst_to_html import rst_to_html  # noqa: F401
 
 try:
     shlex_join = shlex.join  # type: ignore[attr-defined]
@@ -144,23 +134,6 @@ XML = etree.XML
 defaultdict = collections.defaultdict
 
 UNKNOWN = "unknown"
-
-
-class FakeStream:
-    def __init__(self, error):
-        self.__error = error
-
-    log_ = get_logger("docutils")
-
-    def write(self, str):
-        if len(str) > 0 and not str.isspace():
-            if self.__error:
-                raise Exception(str)
-            self.log_.warning(str)
-
-
-docutils_template_path = os.path.join(os.path.dirname(__file__), "docutils_template.txt")
-docutils_writer = docutils.writers.html4css1.Writer()
 
 
 def str_removeprefix(s: str, prefix: str):
@@ -975,49 +948,6 @@ class Params:
 
     def update(self, values):
         self.__dict__.update(values)
-
-
-def get_publisher():
-
-    no_report_level = docutils.utils.Reporter.SEVERE_LEVEL + 1
-    settings_overrides = {
-        "embed_stylesheet": False,
-        "template": docutils_template_path,
-        "warning_stream": FakeStream(False),
-        "doctitle_xform": False,  # without option, very different rendering depending on
-        # number of sections in help content.
-        "halt_level": no_report_level,
-    }
-    # in normal operation we don't want noisy warnings, that's tool author business
-    settings_overrides["report_level"] = no_report_level
-
-    Publisher = docutils.core.Publisher
-    pub = Publisher(
-        parser=None,
-        writer=docutils_writer,
-        settings=None,
-        source_class=docutils.io.StringInput,
-        destination_class=docutils.io.StringOutput,
-    )
-    pub.set_components("standalone", "restructuredtext", "pseudoxml")
-    pub.process_programmatic_settings(None, settings_overrides, None)
-    return pub
-
-
-publisher = get_publisher()
-
-
-@functools.lru_cache(maxsize=10000)
-def rst_to_html(s, error=False):
-    """Convert a blob of reStructuredText to HTML"""
-
-    # if docutils_core is None or docutils_utils is None:
-    #     raise Exception("Attempted to use rst_to_html but docutils unavailable.")
-
-    publisher.set_source(s, None)
-    publisher.set_destination(None, None)
-    output = publisher.publish(enable_exit_status=False)
-    return unicodify(output)
 
 
 def xml_text(root, name=None):
